refactor(sprint-2/b3): merge split chain-integrity test files
tests/security/chain-integrity.test.ts (3 tests, 79 LOC) collapsed into
tests/unit/chain-integrity.test.ts (10 tests, 288 LOC) as a sibling
"security: chain integrity verification" describe block. Both test
suites share imports (appendBlock, verifyChainIntegrity) so co-locating
removes redundancy without losing categorization.

13 tests in the merged file, all green. The only file pair across
tests/{unit,security,integration}/ that shared a name; no other splits
to chase per Sprint 2 scope.
